Lots of things #2

+++ MovieWebsite.Blogs/Controllers/DraftsController.cs
@@ -11,9 +11,25 @@ using MovieWebsite.Shared;
 
 namespace MovieWebsite.Blogs.Controllers
 {
+    /// <summary>
+    /// Contains endpoints for drafts API
+    /// </summary>
     [Route("drafts")]
     public class DraftsController : Controller
     {
+        
+        /// <summary>
+        /// An endpoint to add new draft (name only)
+        /// </summary>
+        /// <param name="title">Name of draft</param>
+        /// <returns>Object with guid of new draft</returns>
+        /// <response code="401">Authorize to use this endpoint</response>
+        /// <response code="400">Drafts limit reached</response>
+        /// <response code="200">Draft successfully added</response>
+        /// <returns>Json with object which contains guid</returns>
+        [ProducesResponseType(typeof(GuidTitlePair), 200)]
+        [ProducesResponseType(401)]
+        [ProducesResponseType(400)]
         [HttpPut("new/{title}")]
         [Authorize]
         public async Task<IActionResult> AddDraft(string title)
@@ -23,6 +39,9 @@ namespace MovieWebsite.Blogs.Controllers
             if (!userId.HasValue) return Unauthorized();
             
             await using var db = new DraftsContext();
+            if (await db.Drafts.Where(x => x.AuthorId == userId).CountAsync() >= 10)
+                return BadRequest("Drafts limit reached");
+            
             var draft = new Post
             {
                 AuthorId = userId.Value,
@@ -33,12 +52,26 @@ namespace MovieWebsite.Blogs.Controllers
             };
             await db.Drafts.AddAsync(draft);
             await db.SaveChangesAsync();
-            return Json(new {guid=draft.Guid});
+            return Json(new GuidTitlePair {Guid=draft.Guid, Title = draft.Title});
         }
         
+        /// <summary>
+        /// An endpoint to update draft
+        /// </summary>
+        /// <param name="guid">Guid of draft to update</param>
+        /// <param name="body">Body of updated draft</param>
+        /// <response code="401">Authorize to use this endpoint</response>
+        /// <response code="403">You have no access to that draft</response>
+        /// <response code="404">Draft not found</response>
+        /// <response code="200">Draft successfully added</response>
+        /// <returns>Nothing</returns>
+        [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
+        [ProducesResponseType(403)]
+        [ProducesResponseType(400)]
         [HttpPut("{guid:Guid}")]
         [Authorize]
-        public async Task<IActionResult> UpdateDraft(Guid guid, [FromBody] UpdatePostBody body)
+        public async Task<IActionResult> UpdateDraft(Guid guid, [FromBody] DraftRequestBody body)
         {
             var identity = HttpContext.User.Identity as ClaimsIdentity;
             var userId = identity.GetIdClaim();
@@ -46,9 +79,9 @@ namespace MovieWebsite.Blogs.Controllers
             
             await using var db = new DraftsContext();
             var draft = await db.Drafts.FirstOrDefaultAsync(x => x.Guid == guid);
-
-            if (body.BodyHtml is not null)
-                draft.BodyHtml = body.BodyHtml;
+            if (draft is null) return NotFound("Draft not found");
+            if (draft.AuthorId != userId) return Forbid("You have no access to that draft");
+            
             if (body.BodyRaw is not null)
                 draft.BodyRaw = body.BodyRaw;
             if (body.Title is not null)
@@ -61,6 +94,17 @@ namespace MovieWebsite.Blogs.Controllers
             return Ok();
         }
         
+        /// <summary>
+        /// An endpoint to delete draft
+        /// </summary>
+        /// <param name="guid">Guid of draft to delete</param>
+        /// <response code="401">Authorize to use this endpoint</response>
+        /// <response code="403">You have no access to that draft</response>
+        /// <response code="404">Draft not found</response>
+        /// <response code="200">Draft successfully deleted</response>
+        /// <returns>Nothing</returns>
+        [ProducesResponseType(200)]
+        [ProducesResponseType(401)]
         [HttpDelete("{guid:Guid}")]
         [Authorize]
         public async Task<IActionResult> DeleteDraft(Guid guid)
@@ -71,12 +115,21 @@ namespace MovieWebsite.Blogs.Controllers
             
             await using var db = new DraftsContext();
             var draft = await db.Drafts.FirstOrDefaultAsync(x => x.Guid == guid);
+            if (draft is null) return NotFound("Draft not found");
             if (draft.AuthorId != userId) return Forbid();
             db.Drafts.Remove(draft);
             await db.SaveChangesAsync();
             return Ok();
         }
         
+        /// <summary>
+        /// An endpoint to get user's drafts
+        /// </summary>
+        /// <response code="401">Authorize to use this endpoint</response>
+        /// <response code="200">Success</response>
+        /// <returns>Array of draft guids</returns>
+        [ProducesResponseType(200, Type = typeof(GuidTitlePair[]))]
+        [ProducesResponseType(401)]
         [HttpGet]
         [Authorize]
         public async Task<IActionResult> FetchDrafts()
@@ -86,10 +139,24 @@ namespace MovieWebsite.Blogs.Controllers
             if (!userId.HasValue) return Unauthorized();
             
             await using var db = new DraftsContext();
-            var draftGuids = await db.Drafts.AsNoTracking().Where(x => x.AuthorId == userId).Select(x=> new {guid=x.Guid, title=x.Title}).ToArrayAsync();
+            var draftGuids = await db.Drafts.AsNoTracking()
+                .Where(x => x.AuthorId == userId)
+                .Select(x=> new GuidTitlePair {Guid = x.Guid, Title = x.Title}).ToArrayAsync();
             return Json(draftGuids);
         }
         
+        /// <summary>
+        /// An endpoint to get draft
+        /// </summary>
+        /// <param name="guid">Guid of draft to get</param>
+        /// <response code="401">Authorize to use this endpoint</response>
+        /// <response code="403">You have no access to that draft</response>
+        /// <response code="404">Draft not found</response>
+        /// <response code="200">Success</response>
+        /// <returns>Json of draft</returns>
+        [ProducesResponseType(200, Type = typeof(ShortPost))]
+        [ProducesResponseType(401)]
+        [ProducesResponseType(403)]
         [HttpGet("{guid:Guid}")]
         [Authorize]
         public async Task<IActionResult> FetchDraft(Guid guid)
@@ -100,16 +167,49 @@ namespace MovieWebsite.Blogs.Controllers
             
             await using var db = new DraftsContext();
             var draft = await db.Drafts.AsNoTracking().FirstOrDefaultAsync(x => x.Guid == guid);
+            if (draft is null) return NotFound("Draft not found");
             if (draft.AuthorId != userId) return Forbid();
-            return Json(draft);
+            return Json(draft.Short);
         }
 
-        public class UpdatePostBody
+        /// <summary>
+        /// Body of draft endpoints
+        /// </summary>
+        public class DraftRequestBody
         {
+            /// <summary>
+            /// Raw json of post for Editor.js
+            /// </summary>
             public string? BodyRaw { get; set; }
+            /// <summary>
+            /// Html body of post
+            /// </summary>
+            [Obsolete("Not used anymore due to vulnerability")]
             public string? BodyHtml { get; set; }
+            /// <summary>
+            /// Title of post
+            /// </summary>
             public string? Title { get; set; }
+            /// <summary>
+            /// Admin's note.
+            /// </summary>
             public string? Note { get; set; }
+        }
+
+        
+        /// <summary>
+        /// Container for guid+title
+        /// </summary>
+        public class GuidTitlePair
+        {
+            /// <summary>
+            /// Value
+            /// </summary>
+            public Guid Guid { get; set; }
+            /// <summary>
+            /// Title
+            /// </summary>
+            public string Title { get; set; }
         }
     }
 }
